Stage 2c: converge the 4e ML cluster (utils + learning), completing #1188
Port the last 2 deferred pairs so there is one module per topic:
- learning.DataSet.sanitize now drops the target (the 4e convention) - verified
  it leaves the 3e learning tests green and is what the neural-net learners need.
- learning.err_ratio / grade_learner accept either a callable predictor or a
  learner object with a .predict method (so the 4e class-learners work through
  the canonical learning); backward-compatible with the 3e closures.
- deep_learning4e imports the canonical aima.utils and defines locally the
  recursive vector helpers (element_wise_product/vector_add/scalar_vector_product)
  the neural net needs on nested weights (aima.utils keeps the flat tuple/np
  versions required by the grid/search code).
- Delete utils4e.py and learning4e.py (no genuinely-unique symbols remained -
  model_selection duplicated cross_validation_wrapper); repoint test_deep_learning4e
  and Learners.ipynb to aima.learning; drop the duplicate test_learning4e.

All 9 X/X4e pairs are now unified. Full suite: 393 passed; docs 0 warnings.
