+++ tests/test_deep_learning4e.py
@@ -2,7 +2,7 @@ import pytest
 from keras.datasets import imdb
 
 from aima.deep_learning4e import *
-from aima.learning4e import DataSet, grade_learner, err_ratio
+from aima.learning import DataSet, grade_learner, err_ratio
 
 random.seed("aima-python")
